Revert "Add archive detail page component"
